fix: apply PR feedback - consolidate autocomplete tests
Addressed:
- Removed duplicate LoginForm autocomplete tests (already existed in main test file)
- Consolidated RegisterForm autocomplete tests into main RegisterForm.test.tsx
- Consolidated AccountTab autocomplete tests into main AccountTab.test.tsx
- Deleted separate .autocomplete.test.tsx files
- All tests passing (1776/1778)
- Following project convention of single test file per component
