--- frontend/src/components/features/auth/__tests__/LoginForm.autocomplete.test.tsx
@@ -1,40 +1,0 @@
-/// <reference types="vitest/globals" />
-import { render, screen } from '@testing-library/react';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { LoginForm } from '../LoginForm';
-import { vi } from 'vitest';
-
-// Mock Next.js router
-vi.mock('next/navigation', () => ({
-  useRouter: vi.fn(() => ({
-    push: vi.fn(),
-    replace: vi.fn(),
-    prefetch: vi.fn(),
-  })),
-}));
-
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: { retry: false },
-  },
-});
-
-const wrapper = ({ children }: { children: React.ReactNode }) => (
-  <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-);
-
-describe('LoginForm - Autocomplete', () => {
-  it('should have autocomplete="current-password" on password field', () => {
-    render(<LoginForm />, { wrapper });
-
-    const passwordInput = screen.getByLabelText(/contraseña/i);
-    expect(passwordInput).toHaveAttribute('autocomplete', 'current-password');
-  });
-
-  it('should have autocomplete="email" on email field', () => {
-    render(<LoginForm />, { wrapper });
-
-    const emailInput = screen.getByLabelText(/email/i);
-    expect(emailInput).toHaveAttribute('autocomplete', 'email');
-  });
-});
